pyproject.toml
[project]
name = "crm-restate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "hypercorn>=0.17.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "restate-sdk[serde]>=0.8.1",
    "uvicorn>=0.35.0",
]
